progress on extract zip
