+++ pkg/report/commons.go
@@ -65,22 +65,6 @@ func getPlatforms(queries model.VulnerableQuerySlice) string {
 	return strings.Join(platforms, ", ")
 }
 
-func getRelativePath(basePath, filePath string) string {
-	var rtn string
-	if strings.Contains(filePath, ".zip") {
-		rtn = filePath
-	} else {
-		relativePath, err := filepath.Rel(basePath, filePath)
-		if err != nil {
-			log.Error().Msgf("Cannot make %s relative to %s", filePath, basePath)
-			rtn = filePath
-		} else {
-			rtn = relativePath
-		}
-	}
-	return rtn
-}
-
 // ExportJSONReport - encodes a given body to a JSON file in a given filepath
 func ExportJSONReport(path, filename string, body interface{}) error {
 	if !strings.Contains(filename, ".") {
