+++ pkg/report/json.go
@@ -1,9 +1,5 @@
 package report
 
-import (
-	"os"
-)
-
 const jsonExtension = ".json"
 
 // PrintJSONReport prints on JSON file the summary results
@@ -14,17 +10,6 @@ func PrintJSONReport(path, filename string, body interface{}) error {
 			return err
 		}
 
-		basePath, err := os.Getwd()
-		if err != nil {
-			return err
-		}
-
-		for i := range summary.Queries {
-			query := summary.Queries[i]
-			for j := range query.Files {
-				query.Files[j].FileName = getRelativePath(basePath, query.Files[j].FileName)
-			}
-		}
 		body = summary
 	}
