+++ pkg/report/model/gitlab_sast.go
@@ -2,13 +2,11 @@ package model
 
 import (
 	"fmt"
-	"path/filepath"
 	"strings"
 	"time"
 
 	"github.com/Checkmarx/kics/internal/constants"
 	"github.com/Checkmarx/kics/pkg/model"
-	"github.com/rs/zerolog/log"
 )
 
 const timeFormat = "2006-01-02T15:04:05" // YYYY-MM-DDTHH:MM:SS a.k.a ISO8601
@@ -112,14 +110,14 @@ func initGitlabSASTScan(start, end time.Time) gitlabSASTScan {
 // BuildGitlabSASTVulnerability adds a new vulnerability struct to vulnerability slice
 func (glsr *gitlabSASTReport) BuildGitlabSASTVulnerability(issue *model.VulnerableQuery, file *model.VulnerableFile) {
 	if len(issue.Files) > 0 {
-		absBasePath, err := filepath.Abs(glsr.basePath)
-		if err != nil {
-			log.Err(err)
-		}
-		relativePath, err := filepath.Rel(absBasePath, file.FileName)
-		if err != nil {
-			log.Err(err)
-		}
+		// absBasePath, err := filepath.Abs(glsr.basePath)
+		// if err != nil {
+		// 	log.Err(err)
+		// }
+		// relativePath, err := filepath.Rel(absBasePath, file.FileName)
+		// if err != nil {
+		// 	log.Err(err)
+		// }
 
 		vulnerability := gitlabSASTVulnerability{
 			ID:       file.SimilarityID,
@@ -138,7 +136,7 @@ func (glsr *gitlabSASTReport) BuildGitlabSASTVulnerability(issue *model.Vulnerab
 				},
 			},
 			Location: gitlabSASTVulnerabilityLocation{
-				File:  relativePath,
+				File:  file.FileName,
 				Start: file.Line,
 				End:   file.Line,
 			},
